requirements.txt
requests
langchain
langchain-community
langchain-openai
pydantic
python-dotenv